Move setInnerHTML to own file
