+++ services/marko-example-website/client/utils/set-inner-html.js
@@ -1,0 +1,10 @@
+export default (ele, html) => {
+  // eslint-disable-next-line no-param-reassign
+  ele.innerHTML = html;
+  Array.from(ele.querySelectorAll('script')).forEach((oldScript) => {
+    const newScript = document.createElement('script');
+    Array.from(oldScript.attributes).forEach(attr => newScript.setAttribute(attr.name, attr.value));
+    newScript.appendChild(document.createTextNode(oldScript.innerHTML));
+    oldScript.parentNode.replaceChild(newScript, oldScript);
+  });
+};
